menu: base code for units properties, but commented out
There are major performance issues with creating a large table of
entries. Until this issue is resolved, through a dedicated widget,
upstream, through a hack or via a function I'm not aware of, units
(and upgrade) properties will not be activated.

Examine callgrind.out (with callgrind_annotate) to profile what is going
wrong in ELM...