snapshot: add snapshotting feature
This is a bit complex. To handle undo/redo and other features
I think of, we are using snapshots: a full image of the map
at a given time.
As one can image, these snapshots would be HEAAAAVY. 256K per
map for a 128x128. So what we do it that we compress in RAM
with the xz algorithm, and manage to reach <10K per snapshot.
Then we navigate in the snapshots for undo/redo.

redo is not implemented yet though...
It might be nice to sync the editor during the snapshot, and to
snapshot the pud itself. But that's for later, and after benchmark...