preview: greatly simplify previewer and fix related bugs
Instead of asking libpud to generate a preview of the map, just use our
own minimap. We grap the pixel data and push it to the preview. So
everything keeps in sync, and we don't need to save the file before
doing a preview. That's WAAAAY better, and I'm very much happy about
this commit. Random self-applause there :)